Move hunk related components into hunk feature folder
